+++ frontend/src/pages/courseDetail.jsx
